Clean up tesseract_rosutils
